requirements.txt
telethon
python-dotenv
psycopg2-binary
dbt-postgres
fastapi
uvicorn
dagster
dagster-webserver
ultralytics
